requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
alembic>=1.11.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
faker>=18.0.0